+++ lib/trpc.ts
@@ -1,6 +1,6 @@
 import { initTRPC, TRPCError } from '@trpc/server';
 import superjson from 'superjson';
-import { Context, createContext } from './context';
+import { Context } from './context';
 import { extendSuperjson } from './extendSuperjson';
 
 extendSuperjson(superjson)
@@ -16,9 +16,9 @@ export const publicProcedure = t.procedure
 
 const isAuth = t.middleware(async ({ next, ctx }) => {
   const token = ctx.req.headers.authorization;
-  // const refreshToken = ctx.req.headers.refreshToken;
+  const refreshToken = ctx.req.headers.refreshToken;
   try {
-    const { data, error } = await ctx.supabase.auth.getUser(token);
+    const { data } = await ctx.supabase.auth.getUser(token);
     if (!data?.user) {
       throw new TRPCError({ code: 'FORBIDDEN', message: 'Not Authenticated' })
     }
