+++ pages/api/trpc/[trpc].ts
@@ -1,6 +1,6 @@
 import * as trpcNext from '@trpc/server/adapters/next';
-import { appRouter } from '../../../backend/app';
-import { createContext } from '../../../lib/context';
+import { appRouter } from '@/backend/app';
+import { createContext } from '@/lib/context';
 // export API handler
 // @see https://trpc.io/docs/api-handler
 export default trpcNext.createNextApiHandler({
